+++ src/web/src/modules/editor/index.ts
@@ -1,6 +1,7 @@
-export { BlogsEditorWorkspace } from "./editor.workspace";
+export { BlogsEditorWorkspace } from "./editor.workspace.js";
+export { configureBlogsEditorClient } from "./editor.services.js";
 export type {
   BlogsEditorHost,
   BlogAuthorOption,
   BlogMediaFile,
-} from "./editor.types";
+} from "./editor.types.js";