+++ src/web/src/modules/public-blog/index.ts
@@ -1,5 +1,5 @@
-export { PublicBlogPage } from "./public-blog.page";
-export { PublicArticlePage } from "./public-article.page";
-export { getPublicArticle, searchPublicArticles } from "./public-blog.services";
-export { blogMediaUrl, blogPlaceholder } from "./media";
-export type { PublicArticle } from "./public-blog.types";
+export { PublicBlogPage } from "./public-blog.page.js";
+export { PublicArticlePage } from "./public-article.page.js";
+export { getPublicArticle, searchPublicArticles } from "./public-blog.services.js";
+export { blogMediaUrl, blogPlaceholder } from "./media.js";
+export type { PublicArticle } from "./public-blog.types.js";